% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_regression.R
\name{anchor_regression}
\alias{anchor_regression}
\title{anchor_regression}
\usage{
anchor_regression(x, anchor, gamma, target_variable)
}
\arguments{
\item{x}{x is a dataframe containing the matrix x}

\item{anchor}{anchor is a dataframe containing the matrix anchor}

\item{gamma}{gamma is the regularization parameter for the Anchor Regression}

\item{target}{target is the target variable name contained in the x dataframe}
}
\value{
A list with coefficient values and a list with the respective names \code{overview_print}
}
\description{
Perform an Anchor Regresseion as described in Rothenhäusler et al.2020
}
\examples{
data(example)
output_table <- anchor_regression(x, anchor, gamma, target_variable)
}
