% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_anchor_regression.R
\name{weighted_anchor_regression}
\alias{weighted_anchor_regression}
\title{weighted_anchor_regression}
\usage{
weighted_anchor_regression(
  data_x_list,
  data_anchor_list,
  gamma,
  target_variable,
  anchor_model_pre = NULL,
  test_split = 0.4
)
}
\arguments{
\item{data_x_list}{list containing anchor dataframes for different environments}

\item{data_anchor_list}{list containing anchor dataframes for different environments}

\item{gamma}{is the regularization parameter for the Anchor Regression}

\item{target_variable}{is the target variable name contained in the x dataframe}

\item{anchor_model_pre}{is the pre estimated model for the Anchor Regression. In case of NULL a new model is estimated.}

\item{test_split}{is desired test/train split for the estimation}
}
\value{
A list estimated coefficients with names and the raw coefficient matrix
}
\description{
Estimates weighted Anchor Regression coefficients
}
